# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solvent API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/pairs": {
      "get": {
        "tags": [
          "crate::http::app::pairs"
        ],
        "operationId": "pairs",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter by token symbol or address",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "wallet",
            "in": "query",
            "description": "Include this wallet's per-side balances",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_List_PairInfo"
                }
              }
            }
          }
        }
      }
    },
    "/v1/pairs/history": {
      "get": {
        "tags": [
          "crate::http::app::pairs"
        ],
        "operationId": "pair_history",
        "parameters": [
          {
            "name": "base",
            "in": "query",
            "description": "Base token address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quote",
            "in": "query",
            "description": "Quote token address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "in": "query",
            "description": "Historical chart window",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PriceHistoryPeriod"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_PairPriceHistory"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/swap": {
      "post": {
        "tags": [
          "crate::http::app::swap"
        ],
        "summary": "Submit a taker-signed order: decode → verify the swapper signature → cosign → route → reserve →\npersist → fill. A malformed or unverifiable order is `400`; an unroutable one returns a\n`declined` trade (`200`).",
        "operationId": "submit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_SwapResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or unverifiable order"
          }
        }
      }
    },
    "/v1/swap/quote": {
      "post": {
        "tags": [
          "crate::http::app::swap"
        ],
        "summary": "Route `amount_in` of `token_in` into `token_out`, returning the split and its impact. `422` when\nno route exists (no makers, or the size is beyond the book).",
        "operationId": "quote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "No route for the pair and size"
          }
        }
      }
    },
...
    "/v1/wallets/{addr}/balances": {
      "get": {
        "tags": [
          "crate::http::app::balances"
        ],
        "summary": "Balances for `addr` over every catalog token. A malformed address → `400`.",
        "operationId": "balances",
        "parameters": [
          {
            "name": "addr",
            "in": "path",
            "description": "Wallet address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_List_TokenBalance"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Amount": {
        "type": "object",
        "description": "A token amount: `raw` base units (never a float), a human-formatted `display`, and `usd` (`None`\nuntil pricing is wired).",
        "required": [
          "raw",
          "display"
        ],
        "properties": {
          "display": {
            "type": "string"
          },
          "raw": {
            "type": "string"
          },
          "usd": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
...
      "PairKind": {
        "type": "string",
        "description": "Whether a pair is stable/stable or involves a volatile asset — drives the wizard's default curve.",
        "enum": [
          "stable",
          "volatile"
        ]
      },
...
      "PairPricePoint": {
        "type": "object",
        "description": "One timestamped pair midpoint, quoted as `quote` per one `base`.",
        "required": [
          "timestamp_ms",
          "price"
        ],
        "properties": {
          "price": {
            "type": "number",
            "format": "double"
          },
          "timestamp_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "volume_usd": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "USD-denominated market volume for the interval when the source provides it."
          }
        }
      },
      "PairWallet": {
        "type": "object",
        "description": "A maker's wallet balance of each side of a pair, as display numbers and exact base units.",
        "required": [
          "base",
          "quote",
          "base_raw",
          "quote_raw"
        ],
        "properties": {
          "base": {
            "type": "number",
            "format": "double"
          },
          "base_raw": {
            "type": "string"
          },
          "quote": {
            "type": "number",
            "format": "double"
          },
          "quote_raw": {
            "type": "string"
          }
        }
      },
...
      "PriceHistoryPeriod": {
        "type": "string",
        "description": "The history window requested by a price chart.",
        "enum": [
          "7d",
          "3m",
          "all"
        ]
      },
...
      "QuoteLeg": {
        "type": "object",
        "description": "One maker's slice of a quoted route.",
        "required": [
          "maker",
          "strategy_hash",
          "token_in",
          "token_out",
          "amount_in",
          "amount_out",
          "curve",
          "share_pct"
        ],
        "properties": {
          "amount_in": {
            "$ref": "#/components/schemas/Amount"
          },
          "amount_out": {
            "$ref": "#/components/schemas/Amount"
          },
          "curve": {
            "type": "string",
            "description": "The curve shape: `XYC` | `Concentrated` | `Pegged`."
          },
          "maker": {
            "type": "string"
          },
          "share_pct": {
            "type": "number",
            "format": "double",
            "description": "This leg's share of the blended output, in percent."
          },
          "strategy_hash": {
            "type": "string"
          },
          "token_in": {
            "$ref": "#/components/schemas/Token"
          },
          "token_out": {
            "$ref": "#/components/schemas/Token"
          }
        }
      },
      "QuoteRequest": {
        "type": "object",
        "description": "A quote request: the pair, and the input size in base units.",
        "required": [
          "token_in",
          "token_out",
          "amount_in"
        ],
        "properties": {
          "amount_in": {
            "type": "string",
            "description": "The input amount, in base units (a decimal integer string)."
          },
          "token_in": {
            "type": "string"
          },
          "token_out": {
            "type": "string"
          }
        }
      },
...
      "Response_List_PairInfo": {
        "type": "object",
        "description": "The envelope wrapping every response. `status_code` sets the HTTP status (never serialized).",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {
            "type": "object",
            "description": "A page of a collection: the items plus an opaque `next_cursor` (absent on the last page) and an\noptional `total`. Carried inside the response envelope's `result`.",
            "required": [
              "items"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "type": "object",
                  "description": "One tradeable pair offered by the Create wizard: the two tokens, their kind, mid price, the\nsuggested defaults, and (when a wallet is supplied) the maker's balance of each side.",
                  "required": [
                    "base",
                    "quote",
                    "type",
                    "default_fee_bps",
                    "default_band_pct"
                  ],
                  "properties": {
                    "base": {
                      "$ref": "#/components/schemas/Token"
                    },
                    "default_band_pct": {
                      "type": "number",
                      "format": "double"
                    },
                    "default_fee_bps": {
                      "type": "integer",
                      "format": "int32",
                      "minimum": 0
                    },
                    "mid": {
                      "type": [
                        "number",
                        "null"
                      ],
                      "format": "double",
                      "description": "Mid price, quote per 1 base — `None` when either side is unpriced."
                    },
                    "quote": {
                      "$ref": "#/components/schemas/Token"
                    },
                    "type": {
                      "$ref": "#/components/schemas/PairKind",
                      "description": "`stable` when both sides are stablecoins, else `volatile`."
                    },
                    "wallet": {
                      "oneOf": [
                        {
                          "type": "null"
                        },
                        {
                          "$ref": "#/components/schemas/PairWallet",
                          "description": "The maker's wallet balance of each side; present only when a wallet was supplied."
                        }
                      ]
                    }
                  }
                }
              },
              "next_cursor": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "total": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              }
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
      "Response_List_TokenBalance": {
        "type": "object",
        "description": "The envelope wrapping every response. `status_code` sets the HTTP status (never serialized).",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {
            "type": "object",
            "description": "A page of a collection: the items plus an opaque `next_cursor` (absent on the last page) and an\noptional `total`. Carried inside the response envelope's `result`.",
            "required": [
              "items"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "type": "object",
                  "description": "One token a wallet holds: its on-chain `balance` and the `pullable` amount a pull could take now\n(`min(balance, allowance→settlement)`).",
                  "required": [
                    "token",
                    "balance",
                    "pullable"
                  ],
                  "properties": {
                    "balance": {
                      "$ref": "#/components/schemas/Amount"
                    },
                    "pullable": {
                      "$ref": "#/components/schemas/Amount"
                    },
                    "token": {
                      "$ref": "#/components/schemas/Token"
                    }
                  }
                }
              },
              "next_cursor": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "total": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              }
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
      "Response_PairPriceHistory": {
        "type": "object",
        "description": "The envelope wrapping every response. `status_code` sets the HTTP status (never serialized).",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {
            "type": "object",
            "description": "Historical market data for one oriented pair.",
            "required": [
              "base",
              "quote",
              "period",
              "points"
            ],
            "properties": {
              "base": {
                "type": "string"
              },
              "period": {
                "$ref": "#/components/schemas/PriceHistoryPeriod"
              },
              "points": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PairPricePoint"
                }
              },
              "quote": {
                "type": "string"
              }
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
      "Response_QuoteResponse": {
        "type": "object",
        "description": "The envelope wrapping every response. `status_code` sets the HTTP status (never serialized).",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {
            "type": "object",
            "description": "A read-only quote: the routed split, its blended output, and price impact. `quote_id` is\ndeterministic from the request; `expires_at` is an advisory RFC-3339 instant (a quote does not\nbind a future swap).",
            "required": [
              "quote_id",
              "amount_out",
              "price_impact_pct",
              "makers_sourced",
              "legs",
              "expires_at"
            ],
            "properties": {
              "amount_out": {
                "$ref": "#/components/schemas/Amount"
              },
              "expires_at": {
                "type": "string"
              },
              "legs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/QuoteLeg"
                }
              },
              "makers_sourced": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "price_impact_pct": {
                "type": "number",
                "format": "double"
              },
              "quote_id": {
                "type": "string"
              }
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
      "Response_SwapResponse": {
        "type": "object",
        "description": "The envelope wrapping every response. `status_code` sets the HTTP status (never serialized).",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {
            "type": "object",
            "description": "The created trade and where its lifecycle stands (`created`…`submitted`, or `declined`).",
            "required": [
              "trade_id",
              "status"
            ],
            "properties": {
              "status": {
                "type": "string"
              },
              "trade_id": {
                "type": "string"
              }
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
      "Status": {
        "type": "string",
        "description": "Whether an API call succeeded.",
        "enum": [
          "Ok",
          "Error"
        ]
      },
...
      "SwapRequest": {
        "type": "object",
        "description": "A taker-signed order submission, mirroring the UniswapX Orders API (`{ encodedOrder, signature,\nchainId, quoteId? }`). The taker's client builds + signs the base order; the server cosigns.",
        "required": [
          "encodedOrder",
          "signature",
          "chainId"
        ],
        "properties": {
          "chainId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "encodedOrder": {
            "type": "string"
          },
          "quoteId": {
            "type": [
              "string",
              "null"
            ]
          },
          "signature": {
            "type": "string"
          }
        }
      },
...
      "Token": {
        "type": "object",
        "description": "A token's identity + display essentials — embedded wherever a response references a token\nwithout the full [`Asset`] (pool base/quote, trade flows, roster).",
        "required": [
          "address",
          "chain_id",
          "symbol",
          "decimals"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "chain_id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "decimals": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "symbol": {
            "type": "string"
          }
        }
      },
...
    }
  }
}